dentweb-auto/requirements.txt
pywinauto==0.6.8
pyautogui==0.9.54
PyYAML==6.0.2
requests==2.32.3
python-dotenv==1.0.1
